+++ src/main/java/codesquad/application/config/DatabaseConfig.java
@@ -3,9 +3,10 @@ package codesquad.application.config;
 import codesquad.application.bean.Bean;
 import codesquad.application.database.ArticleDatabase;
 import codesquad.application.database.ArticleMemoryDatabase;
+import codesquad.application.database.SessionStorage;
 import codesquad.application.database.UserDatabase;
 import codesquad.application.database.UserMemoryDatabase;
-import codesquad.application.database.SessionStorage;
+import codesquad.application.database.SessionMemoryStorage;
 
 public class DatabaseConfig {
 
@@ -15,8 +16,8 @@ public class DatabaseConfig {
     }
 
     @Bean
-    public SessionStorage userSessionStorage() {
-        return new SessionStorage();
+    public SessionStorage sessionStorage() {
+        return new SessionMemoryStorage();
     }
 
     @Bean
